Refactors time variables to `uint32_t`
Converts `unsigned long` and `u_int_4` time-related variables to `uint32_t` for consistency and potential memory savings. This change ensures that time values are handled uniformly across the codebase, and that they fit within the expected range, avoiding potential overflow issues.

+++ PAC/common/PAC_info.cpp
@@ -59,9 +59,9 @@ void PAC_info::eval()
         up_mins = up_msec / ( 1000UL * 60 ) % 60;
         up_secs = up_msec / 1000 % 60;
 
-        sprintf( up_time_str, "%lu дн. %02lu:%02lu:%02lu",
-            (u_long)up_days, (u_long)up_hours,
-            (u_long)up_mins, (u_long)up_secs );
+        sprintf( up_time_str,
+            "%" PRIu32 " дн. %02" PRIu32 ":%02" PRIu32 ":%02"  PRIu32,
+            up_days, up_hours, up_mins, up_secs );
         }
     }
 //-----------------------------------------------------------------------------
