+++ PAC_control_projects/main_control_prg/main_cycle.cpp
@@ -19,12 +19,12 @@ extern bool G_READ_ONLY_IO_NODES;
 int main_cycle()
     {
 #ifdef TEST_SPEED
-    static u_long st_time;
+    static uint32_t st_time;
     st_time = get_millisec();
 
-    static u_long all_time = 0;
+    static uint32_t all_time = 0;
 
-    static u_long cycles_cnt = 0;
+    static uint32_t cycles_cnt = 0;
     cycles_cnt++;
 #endif // TEST_SPEED
